module-5/hw-9/gradio/requirements.txt
torch
torchvision
pytorch-lightning
Pillow
gradio
pytest
tokenizers>=0.11.1,<0.14